requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
diskcache==5.6.3
distro==1.9.0
docker==7.0.0
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
openai==1.20.0
packaging==24.0
pillow==10.3.0
proglog==0.1.10
pyautogen==0.2.26
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
termcolor==2.4.0
tiktoken==0.6.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
whisper==1.1.10
